==> S5_Process_mgf.ipynb <==
# Format and merge spectra energies together

Read and write the MGF format with pyteomics and matchms packages.

Pyteomics is easier to use if we just want to clean or modify rapidly the mgf. 
Matchms is good when dealing with the intensities, it has lots of method for spectra filtering and normalization.

from pyteomics import mgf, auxiliary
import numpy as np
from collections import defaultdict
from tqdm import tqdm

# Load the mgf
infile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.mgf"

Structure view of the current MGF

with mgf.read(infile) as reader: 
    auxiliary.print_tree(next(reader))

# Go through the spectra
with mgf.read(infile) as reader: 
    for i, spectrum in enumerate(reader):
        if i > 0: break
        print(i, spectrum)
        #spectrum["params"]["title"] = spectrum["params"]["title"]
        print(spectrum["m/z array"], spectrum['intensity array'])
        print(dict(zip(spectrum["m/z array"], spectrum['intensity array'])))
        

### Step for merging the energies together

from collections import defaultdict
from tqdm import tqdm

modified_spectra = defaultdict(list)
all_mz_intensity_dict = defaultdict(dict)

# input MGF 
infile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.mgf"

with mgf.read(infile, convert_arrays = 1, read_charges = False) as infile:
    for spectrum in tqdm(infile):
        # careful when splitting, inchis have ";" in them... 
        try:
            id, energy, type, cfmid_version, *inchi = spectrum["params"]["title"].split(";")
        except ValueError:
            print(spectrum)
        spectrum["params"]["title"] = id
        spectrum["params"]["cfmid"] = cfmid_version
        spectrum["params"]["InChI"] = ';'.join(inchi)[:-1]
        spectrum["params"]["name"] = id
        mz_array = spectrum["m/z array"]
        intensity_array = spectrum['intensity array']
        current_mz_intensity_dict = dict(zip(mz_array, intensity_array))
    
        if id not in modified_spectra.keys():
            #print(modified_spectra.keys())
            #print("Added new spectra for", id, "energy", energy)
            #print(current_mz_intensity_dict)
            modified_spectra[id] = spectrum
            all_mz_intensity_dict[id] = current_mz_intensity_dict
        else:
            #print("Merged intensities for spectra", id, "energy", energy)
            all_mz_intensity_dict[id] = dict(sorted((*all_mz_intensity_dict[id].items(), *current_mz_intensity_dict.items())))
            modified_spectra[id]["m/z array"] = list(all_mz_intensity_dict[id].keys())
            modified_spectra[id]["intensity array"] = list(all_mz_intensity_dict[id].values())
            #print(*all_mz_intensity_dict[id].items())
            #print(all_mz_intensity_dict[id])

print(len(modified_spectra.keys()))

outfile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.mgf"
!rm -f $outfile

# mgf.write always appends to the file, so we need to remove it if we run it again
mgf.write(spectra=modified_spectra.values(), output=outfile)

# Check the output file
!head $outfile
!grep -c "TITLE" $outfile

#### Convert spectra to GNPS format and create the batchupload tsv file

In order to send our data to GNPS, we have to reformat our MGF like this:

- https://gnps.ucsd.edu/ProteoSAFe/result.jsp?task=1ad7bc366aef45ce81d2dfcca0a9a5e7&view=download_clustered_spectra

And add a tsv file as they show on this link:

- https://ccms-ucsd.github.io/GNPSDocumentation/batchupload/


import sqlite3
import pandas as pd

DB_PATH = "../../Database/compounds.sqlite"

con = sqlite3.connect(DB_PATH)
cur = con.cursor()

compounds_info = pd.read_sql_query('SELECT ID, SMILES FROM compounds', con)

# We need to add the smile code, so we extract it from our database 
# And use a dict consisting only of {ID: SMILES} to retrieve them fast based on ID
series = compounds_info["SMILES"]
series.index = compounds_info["ID"]
compounds_dict = series.to_dict()

from dataclasses import dataclass

@dataclass
class Entry:
    filename: str
    compound_name: str
    moleculemass: float
    extractscan: int
    smiles: str
    inchi: str
    seq: str = "*..*"
    instrument: str = "Hybrid F"
    ionsource: str = "LC-ESI"
    inchiaux: str = "N/A"
    charge: int = 1
    ionmode: str = "Positive"
    pubmed: str = "N/A"
    acquisition: str = "Other"
    exactmass: float = 0
    datacollector: str = "D. Elser, D. Pflieger"
    adduct: str = "M+H"
    interest: str = "N/A"
    libquality: int = 3
    genus: str = "N/A"
    species: str = "N/A"
    strain: str = "N/A"
    casnumber: str = "N/A"
    pi: str = "E. Gaquerel"

    def fields(self):
        fields = [self.filename, self.seq, self.compound_name, self.moleculemass, self.instrument, self.ionsource, self.extractscan, self.smiles, 
            self.inchi, self.inchiaux, self.charge, self.ionmode, self.pubmed, self.acquisition, self.exactmass, self.datacollector, 
             self.adduct, self.interest, self.libquality, self.genus, self.species, self.strain, self.casnumber, self.pi]
        return fields  
        

import csv

infile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.mgf"
outfile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.mgf"
outtsv = '../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.tsv'

modified_spectra = defaultdict(list)

with mgf.read(infile, convert_arrays = 1, read_charges = False) as infile:
    with open(outtsv, "w") as out:
        # Required fields
        header = ["FILENAME", "SEQ", "COMPOUND_NAME", "MOLECULEMASS", "INSTRUMENT", "IONSOURCE", "EXTRACTSCAN", 
                  "SMILES", "INCHI", "INCHIAUX", "CHARGE", "IONMODE", "PUBMED", "ACQUISITION", "EXACTMASS", "DATACOLLECTOR", 
                  "ADDUCT", "INTEREST", "LIBQUALITY", "GENUS", "SPECIES", "STRAIN", "CASNUMBER", "PI"]
        # CSV writer
        writer = csv.writer(out, delimiter='\t')
        writer.writerow(header)
        
        # Loop through the spectra
        for index, spectrum in enumerate(tqdm(infile), start = 1):
            id = spectrum["params"]["title"]
            smile = compounds_dict[id]
            
            Compound = Entry(filename = "TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.mgf", 
                             compound_name = id, 
                             moleculemass = spectrum["params"]["pepmass"][0], 
                             extractscan = index, 
                             smiles = smile, 
                             inchi = spectrum["params"]["inchi"])
                             
            writer.writerow(Compound.fields())

            spectrum["params"]["title"] = f"Scan Number: {index}"
            spectrum["params"]["filename"] = Compound.filename
            spectrum["params"]["scans"] = index
            del spectrum["params"]["inchi"]

            modified_spectra[id] = spectrum

print("Saving MGF now...")
mgf.write(spectra=modified_spectra.values(), output=outfile)
print("MGF was succesfully saved!")

### Using matchms 

from matchms.importing import load_from_mgf
from matchms.exporting import save_as_mgf

infile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.mgf"
outfile = "../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.matchms.mgf"
outcsv = '../../Ressources/CFMID/MGF/TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.matchms.csv'

modified_spectra = []
spectrums = load_from_mgf(infile)

with open(outcsv, "w") as out:
    header = ["FILENAME", "SEQ", "COMPOUND_NAME", "MOLECULEMASS", "INSTRUMENT", "IONSOURCE", "EXTRACTSCAN", 
              "SMILES", "INCHI", "INCHIAUX", "CHARGE", "IONMODE", "PUBMED", "ACQUISITION", "EXACTMASS", "DATACOLLECTOR", 
              "ADDUCT", "INTEREST", "LIBQUALITY", "GENUS", "SPECIES", "STRAIN", "CASNUMBER", "PI"]
    writer = csv.writer(out)
    writer.writerow(header)
    
    for index, spectrum in enumerate(tqdm(spectrums), start = 1):
        id = spectrum.get("title")
        smile = compounds_dict[id]

        Compound = Entry(filename = "TOTAL_COMPOUNDS_DB.energies_merged_name.GNPS_like.mgf", 
                         compound_name = id, 
                         moleculemass = spectrum.get("pepmass")[0], 
                         extractscan = index, 
                         smiles = smile, 
                         inchi = spectrum.get("inchi"))

        writer.writerow(Compound.fields())

        spectrum.set("title", f"Scan Number: {index}")
        spectrum.set("filename",  Compound.filename)
        spectrum.set("scans", index)
        #spectrum.set("smiles", smiles)
                         
        modified_spectra.append(spectrum)

save_as_mgf(spectrums = modified_spectra, filename = outfile)                     


## Conversion from JDX to MGF format

QCxMS saves the spectrum as a .jdx (JCAMP-DX format). 
The output format can not be read directly by the python jdx parser of the jcamp package 

We have to change one problematic line in the all QCxMS jdx file: 

```bash
find . -name *.jdx | while read i; do sed -i 's/##PEAK TABLE=(XY..XY) 1/##PEAK TABLE=(XY..XY)/g' $i; done
```

Load the InChIs of our compounds

import pandas as pd 

jassbi_structure = "/Volumes/6TB/Projects/Gaquerel/Elser/jassbi_structures.txt"

df = pd.read_table(jassbi_structure, delimiter="\t", header = None, names = ["ID", "InChI"])

# check inchi of a compound
df.loc[df.ID == "nicotpanoid_A_"]["InChI"].values.item()

Convert the QCxMS .jdx file to an MGF file with InChI information. 

from jcamp import JCAMP_reader
from matchms import Spectrum
from matchms.exporting import save_as_mgf
from pathlib import Path
from tqdm import tqdm


spectra = []

output_mgf = "/Volumes/6TB/Projects/Gaquerel/Elser/Compounds_databank/Ressources/qcxms_jassbi.mgf"
jassbi_folder = "/Volumes/6TB/Projects/Gaquerel/Elser/Compounds_databank/Ressources/Jassbi_QCxMS/"

jdx_files = Path(jassbi_folder).glob("*/*.jdx")

for jdx in jdx_files: 
    # Get compound name from parent dir 
    compound_name = jdx.parent.name
    # Read the .jdx 
    jdx_dict = JCAMP_reader(jdx)
    
    print(compound_name)
    inchi = df.loc[df.ID == compound_name]["InChI"].values.item()
    
    # Create the spectrum from the jdx_dict informations
    spectrum = Spectrum(mz=jdx_dict["x"],
                        intensities=jdx_dict["y"],
                        metadata={"title": jdx_dict["title"],
                                  "charge": 0,
                                  "inchi": inchi,
                                  "name": compound_name})
    
    spectra.append(spectrum)

save_as_mgf(spectrums = spectra, filename = output_mgf)                     


Check the spectrum of a .jdx file

jcamp_dict = JCAMP_reader(jdx_filename)

import matplotlib.pyplot as plt
from numpy import arange, alen

plt.figure()
for n in arange(alen(jcamp_dict['x'])):
    plt.plot((jcamp_dict['x'][n],jcamp_dict['x'][n]), 
             (0.0, jcamp_dict['y'][n]), 'm-', linewidth=2.0)
plt.title(jdx_filename)
plt.xlabel(jcamp_dict['xunits'])
plt.ylabel(jcamp_dict['yunits'])
