==> S11_mgf-rem-redundancy-v4.ipynb <==
First we define the functions needed and then parse the mgf and csv file 

import sys
import os.path
from pathlib import Path   
from tkinter import * 

from tkinter.filedialog import askopenfilename
from tkinter.filedialog import asksaveasfile

import numpy as np


def tolerance(valuesample,valuedb,ppm):  #calculates the ppm range
    valuemin=valuesample-valuesample*(ppm*10**-6)
    valuemax=valuesample+valuesample*(ppm*10**-6)   
    #print valuemax,valuesample,valuemin,valuedb
    if valuedb<=valuemax and valuedb>=valuemin:
        return True
    else:
        return False
    

cv = lambda x: np.std(x, ddof=1) / np.mean(x) * 100  
    

dialog = Tk()
dialog.title("Import")
fname1 = Path(askopenfilename(filetypes = [("Mzmine GNPS Export", "*.mgf")]))  #deactivate line to remove file dialog
dialog.destroy()

dialog = Tk()
dialog.title("Import")
fname2 = Path(askopenfilename(filetypes = [("Quant Table Export", "*.csv")]))  #deactivate line to remove file dialog
dialog.destroy()
   
#fname1= "2501.mgf"  # enter file name or path activate when file dialog not needed
#fname2="2501_quant.csv" # enter file name or path activate when file dialog not needed

quant=[]

with open(fname2, 'r') as f:     #Reads the Quant table into a List
 #counter=0
 for line in f:
        line=line.strip('\n')
        line1=line.split(',')
        rm= len(line1)-1
        line1.pop(rm)
        quant.append(line1)

print(quant[0])


SampleList=[]

ScanList=[]
with open(fname1, 'r') as f:     #Reads the Samples into a List
 counter=0
 for line in f:
     line=line.strip('\n')
     if  line.startswith('FEATURE_ID'):
         ListID=line.split('=')
         ScanList.append(ListID)       
          
     elif line.startswith('PEPMASS'):
         ListPEPMASS=line.split('=')
         #ListPEPfloat= float(ListPEPMASS[1])
         ScanList.append(ListPEPMASS) 
         
     elif line.startswith('SCANS'):
         ListSCANS=line.split('=')
         ScanList.append(ListSCANS)
         
     elif line.startswith('RTINSECONDS'):
         ListRT=line.split('=')
         ScanList.append(ListRT)
         
     elif line.startswith('CHARGE'):
         ListCHARGE=line.split('=')
         #ListCHARGEfloat= float(ListCHARGE[1])
         ScanList.append(ListCHARGE)
         #ScanList.append([])
         #print ListCHARGE
     elif line.startswith('MSLEVEL'):
         ListRT=line.split('=')
         ScanList.append(ListRT)
         
     elif line.startswith(('0', '1', '2', '3', '4', '5', '6', '7', '8', '9')):
         ListMZandInt=line.split(' ')
         #print(ListMZandInt)
         #ListMZandIntfloat0= float(ListMZandInt[0])
         #ListMZandIntfloat1= float(ListMZandInt[1])
         #ListMZandIntfloat=[ListMZandIntfloat1,ListMZandIntfloat0]
         
         #print(ScanList)
         ScanList.append(ListMZandInt)
         
         #print ScanList
         #sys.exit(0)
         #print ListMZandInt
     elif line.startswith('END IONS'):
         if len(ScanList[5:])>4:   # parse only if more than three fragments
                SampleList.append(ScanList)
         ScanList=[]
            


print(quant[1][3:])
cache=[]
quantcalc=[]
for i in quant:
    cache=[i[0],i[3:]]
    quantcalc.append(cache)
    cache=[]

print(quantcalc[1])


Look at the List that we just parsed



quantcalc.pop(0) #remove header


       
quantcalc1=[]
cache=[]
cacheminmax=[]
for i in quantcalc:
    mi=min(list(np.array(i[1]).astype(float)))  #calculate min and max of intensity across samples
    ma=max(list(np.array(i[1]).astype(float)))
    cacheminmax=[mi,ma]
    cache=[i[0],cacheminmax]
    quantcalc1.append(cache)
    cacheminmax=[]
    cache=[]

   
  
    
print(SampleList[0])
print(SampleList[0][0])
print(SampleList[0][1])
print(SampleList[0][3])
print(SampleList[0][0][1]) #ID
print(SampleList[0][1][1]) #MZ
print(SampleList[0][3][1]) #RT
print(len(SampleList))

print(quantcalc1[135])
print(len(quantcalc1))
print(SampleList[2][6:])
print(len(SampleList[2][6:]))


Take out the MZ's of the List remove duplicates with and then check the duplicates free list with the SampleList with a deviation of 3 ppm, saves RT and ID of the matching features into a new List


mzlist=[]
for i in SampleList:
    mzlist.append(i[1][1]) # Creates a Mz only List

#print(mzlist) 
print(len(mzlist))
mzlistshort=list(set(mzlist))
#print(mzlistshort) 
print(len(mzlistshort))  # Remove duplicates

ID=[]
RT=[]
Int=[]
cachecalc=[]
calc=[]
for a in mzlistshort:
    counter=0 
    for b in SampleList:  
        if tolerance(float(a),float(b[1][1]),3) == True: # check with 3 ppm deviation
            counter+=1
            ID.append(b[0][1])
            RT.append(float((b[3][1])))
            
            for c in quantcalc1:
                if c[0] == b[0][1]:
                    Int.append(c[1])
                else:
                    pass
            cachecalc=[ID,RT,Int]
        else:
            pass
    #countscache=[counter,a]    
    calc.append(cachecalc)
    #calc.append(RT)
    RT=[]
    ID=[]
    Int=[]
    cachecalc=[]
print(calc[0])    

Here we use our List to count the number of ocurrences for each identical mz group and continue only with those that appear more than 5 times, from these we use the RT to calculate the CV and then take only those with a CV greater than 10 % into the final exclucion list.

#print(calc[0:100]) 
print(calc[75])
print(calc[75][1])
print(len(calc[75][1]))
check=[]
for i in calc:
    if len(i[1]) > 5:  # only take features that appear more than five times
        check.append(i)
        #print (i)
    else:
        pass
    
print(len(check))  

print(check[1])
cv(check[1][1])

minlist=[]
maxlist=[]
checkf=[]
cache=[]
for a in check:
    for i in a[2]:
        #print(i[0]) #min
        minlist.append(i[0])
        #print(i[1])# max
        maxlist.append(i[1])
    cache=[a[0],a[1],minlist,maxlist]    
    checkf.append(cache)
    minlist=[]
    maxlist=[]
    cache=[]


excllist=[]
for a in checkf:
    x=[max(a[2]),min(a[3])]
    if cv(a[1])>10 and cv(x)<30:   # calculate the CV based on the RT only take features into the exclusion list with a CV greater than 10 %
        for b in a[0]:             # also calculate  the min of the max and the max of the min across smiliar mz's than take only features with a CV below 30 %
            excllist.append(b)
    else:
        pass
    
print (excllist[0:30])

Here we just again remove duplicates from our exclusion list

print(len(excllist))
excllist=list(set(excllist)) # Remove Duplicates
print(len(excllist)) # Number of featues that will be removed
print(excllist[0:10])


In the next step we remove the features stored in our exclusion list

print(len(SampleList)) #Number of Features before
print(len(quant))
Samplelistshort=[]

for item in SampleList:
        if item[0][1] not in excllist:
            Samplelistshort.append(item)
        else:
            pass
ids_end=[]        
for i in Samplelistshort:
    ids_end.append(i[0][1])
            
        
removed=[]
for item  in quant:
    if item[0] not in ids_end:
            removed.append(item)
    else:
        pass    

quantshort=[quant[0]]

for item  in quant:
        if item[0] in ids_end:
            quantshort.append(item)
        else:
            pass



      
print(len(Samplelistshort)) # Number of Features after removal    
print(len(removed)) # Number of removed Features 
print(len(quantshort)) # Number of Features after removal

print(Samplelistshort[0][0][1])



#print(ids_end)


Here we just write our cleaned Data into a new mgf file and also create the new csv tables



#deactivate when file dialog not needed

dialog = Tk()
dialog.title("Save")
f =asksaveasfile(mode="w",filetypes=[("MGF-format", "*.mgf")])
dialog.destroy()

for i in Samplelistshort:
    #print (i)
    f.write("BEGIN IONS")
    f.write('\n')
    for subi in i:
        #f.write('\n')
        f.write(subi[0])
        if subi[0].startswith(('0', '1', '2', '3', '4', '5', '6', '7', '8', '9')):
            f.write(" ")
        else:
            f.write("=")
        f.write(subi[1])    
        f.write('\n')
        #print(subi)
        f#or subsubi in subi:
            #f.write(subsubi)
            #f.write('#')
    #f.write('\n')        
    f.write("END IONS")
    f.write('\n') 
    f.write('\n')
            
        
f.close()


dialog = Tk()
dialog.title("Save")
f =asksaveasfile(mode="w",filetypes=[("Quant-Table-cleaned", "*.csv")])
dialog.destroy()

for i in quantshort:
    for subi in i:
        f.write(subi)
        f.write(",")
    f.write('\n')
f.close()

dialog = Tk()
dialog.title("Save")
f =asksaveasfile(mode="w",filetypes=[("Removed-features", "*.csv")])
dialog.destroy()    
    
for i in removed:
    for subi in i:
        f.write(subi)
        f.write(",")
    f.write('\n')     
f.close()
#activate when file dialog not needed
"""

with open("/Users/delser/Desktop/PhD/Scripts/list_clean.mgf", 'w') as f:
    for i in SampleList:
        f.write("BEGIN IONS")
        f.write('\n')
        for subi in i:
            f.write(subi[0])
            if subi[0].startswith(('0', '1', '2', '3', '4', '5', '6', '7', '8', '9')):
                f.write(" ")
            else:
                f.write("=")
            f.write(subi[1])
            f.write('\n')
        f.write("END IONS")
        f.write('\n') 
        f.write('\n')

with open("/Users/delser/Desktop/PhD/Scripts/list_clean.csv", 'w') as f:
    for i in quant:
        for subi in i:
            f.write(subi)
            f.write(",")
        f.write('\n')
        
with open("/Users/delser/Desktop/PhD/Scripts/list_removed.csv", 'w') as f:
    for i in removed:
        for subi in i:
            f.write(subi)
            f.write(",")
        f.write('\n')        
"""
