==> S14_degree-unsaturation-sirius.ipynb <==
Script to load Sirius formula_identifications.tsv and restore the feature id, also calculate the degree of unsaturation from the molecular formula
requires molmass package!


from molmass import Formula

from pathlib import Path   
from tkinter import * 

from tkinter.filedialog import askopenfilename
from tkinter.filedialog import asksaveasfile
import csv


def degree(form): # function to calculate degree of unsaturation
    f= Formula(form)
    MF=(f._elements)
    try:
        b=MF["H"][0]
    except:
        b=0
    try:
        c=MF["N"][0]
    except:
        c=0
    try:
        a=MF["C"][0]
    except:
        a=0
    if a ==0 or b==0:
        return("NA")
    else:
        return((a+1)-((b-c)/2))
    
 


dialog = Tk()
dialog.title("Import")
fname = Path(askopenfilename(filetypes = [("Sirius formula_identifications", "*.tsv")]))  # file dialog
dialog.destroy()




Data=[]

with open( fname,'r') as f:  # loads file into ScanList
 for line in f:
  line=line.strip('\n') 
  Data.append (line.split('\t'))
    
print(Data[1])    

print(Data[1][1]) 
print(Data[1][15]) 
print(Data[1][15].split('_'))

head=Data[0]
Data.pop(0)

newdata=[]
for i in Data:
    x=degree(i[1])
    ID=i[15].split('_')
    i.append(ID[2]) # position in ID split
    i.append(x)
    newdata.append(i)
    
print(newdata[0])    
 

head.append("ID")
head.append("degree")

print(head[0])

dialog = Tk()
dialog.title("Save")
csvfile =asksaveasfile(mode="w",filetypes=[("formula_identifications-extended", "*.tsv")]) # file dialog
dialog.destroy()

filewriter = csv.writer(csvfile, delimiter='\t', quotechar='|', quoting=csv.QUOTE_MINIMAL)
filewriter.writerow([head[0],head[1],head[2],head[3],head[4],head[5],head[6],head[7],head[8],head[9],head[10], head[11],head[12],head[13],head[14],head[15],head[16],head[17]])
for item in newdata:
    filewriter.writerow([item[0],item[1],item[2],item[3],item[4],item[5],item[6],item[7],item[8],item[9],item[10], item[11],item[12],item[13],item[14],item[15],item[16],item[17]])
