==> S17_MSLDAmerge-motfs.ipynb <==
import pandas as pd


df_quant = pd.read_csv("/Users/delser/Desktop/PhD/Phytochemistry/FBMN/alltissues/FEATURE-BASED-MOLECULAR-NETWORKING-cf822b6c-15072021/alltissues15072021-quant-py-dil.csv",index_col='row ID', sep=",")

df_motif = pd.read_csv("/Users/delser/Desktop/PhD/Phytochemistry/FBMN/alltissues/FEATURE-BASED-MOLECULAR-NETWORKING-cf822b6c-15072021/MS2LDA-all.csv",index_col='row ID', sep=",")

df_frag=pd.read_csv("/Users/delser/Desktop/PhD/Phytochemistry/FBMN/alltissues/FEATURE-BASED-MOLECULAR-NETWORKING-cf822b6c-15072021/MS2LDA-fragments.csv",index_col='Motif', sep=",")

df_frag.head()


df_motif.head()
df_motif_red=df_motif.drop(["Precursor Mass","Retention Time","Document Annotation","Document","Probability","Overlap Score"], axis=1)

df_motif_red.head()

#result= pd.concat([df_quant, df_motif.reindex(df_quant.index)], axis=1)
#result = pd.concat([df_quant, df_motif], ignore_index=True, sort=False)
result = df_quant.join(df_motif, how="outer")
#result = left.join(right, how="outer")
result.head()

clean=result.drop(["Precursor Mass","Retention Time","Document Annotation","Document"], axis=1)

#clean=clean.reset_index()

#full=result = df_quant.join(df_motif, how="outer")

#clean=clean.set_index("Motif")
#clean.head()

#full = clean.join(df_frag, how="outer")
#full.head()

#full.to_csv("alltissues27042021_all.csv",sep=',')

#occur = full.groupby(['Motif', 'row ID']).size()

#occur.head()

#occur.to_csv("alltissues27042021_counts.csv",sep=',')

occur1 = clean.groupby(['Motif', 'row ID']).size()

occur1.head()
clean.head()
#df.groupby(by=df.columns, axis=1).sum()

new=pd.get_dummies(df_motif_red, columns=['Motif'])
new.head()
#new.to_csv("counts-motif.csv",sep=',')

#df_bool=new.drop_duplicates(subset=["row m/z"])
#df_bool.head()
#df_bool.to_csv("counts-motif-dupl.csv",sep=',')

#x=new.groupby(level=0)['row ID'].sum().reset_index()
x=new.groupby(new.index).sum()
x.head()

#y=x.transpose()
a=df_quant.drop(["row m/z","row retention time"], axis=1)
#a=a.transpose()

#y.head()

#a.head()
x.head()
#x.to_csv("counts-motif.csv",sep=',')

#esult_bool = df_quant.join(x, how="outer")
#result_bool.head()
a.head()

quant0=a.where(a >10000,0) 
quant0.head()

quant1=quant0.where(quant0==0.0,1)
quant1.head()

quant1.to_csv("feature-bin.csv",sep=',')
#nrep=quant1[["07 Nrep"]]
#nrep.head()

#rslt_df = nrep[nrep["07 Nrep"] > 0]
#rslt_df.head()

#test = rslt_df.join(x, how="inner")
#test.head()

#allsum=test.sum(axis=0)

#allsum.head()

headers=quant1.columns.values.tolist()
listindex=x.columns.values.tolist()
#fullindex=headers+listindex
print(headers)

kwargs = {}
for i in headers:
    fsamples=quant1[[i]] #retrieve one sample
    fsamples = fsamples[fsamples[i] > 0] # remove zero occurences of features
    fmotifs = fsamples.join(x, how="inner") # merge with motifs, but only the ones present
    allsums=fmotifs.sum(axis=0)
    kwargs[i] = allsums

#print(kwargs)    

#type(kwargs['07 Nrep'])

df3 = pd.DataFrame(kwargs, index=listindex )

df4 =pd.DataFrame(kwargs, index=headers )
"""
df2 = pd.DataFrame(columns=headers)
for i in headers:
    df1=kwargs[i].to_frame()
    df2.join(df1, how="outer")
    
df2.head()
"""


#liste=["a","b","c"]
#df3 = pd.DataFrame( index=listindex )
#df(3) = pd.DataFrame(columns = ['Name', 'Articles', 'Improved'], 
                   #index = ['a', 'b', 'c'])
#df3.columns=liste
df3.head()

df3.to_csv("alltissues15072021_counts-motif.csv",sep=',')

df4.fillna(0,inplace=True)
ds=df4.sum(axis=0)
ds.to_csv("alltissues15072021_counts-samples.csv",sep=',')
