==> S20_phylometabo.ipynb <==
import pandas as pd
import numpy as np
from scipy.spatial.distance import squareform
from scipy.spatial.distance import pdist

df= pd.read_csv("/Users/delser/Git/phd/input_pairwise_ntw.txt",index_col='cluster', sep="\t")

df=df.transpose()
header=df.columns.values.tolist()
df

pairwise = pd.DataFrame(
    squareform(pdist(df)),
    columns = df.index,
    index = df.index
)

pairwise

pairwise.to_csv("ntw_pairwise.tsv", sep="\t")
