==> S21_phylo.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Cmd+Shift+Enter*. 

```{r}
library(ape)

```


```{r}
x=read.csv("/Users/delser/Git/phd/ntw_pairwise.tsv",sep="\t")
row.names(x) <- x$X
head(x)
z=x[-1]
head(z)

```
```{r}
```


```{r}
y=data.matrix(z)


head.matrix(y)

```

```{r}
#y <- 1-y
y2 <- as.dist(y)
#my_nj=fastme.bal(y2)
my_nj <- ape::nj(y2)
fun<- function(xx) nj(as.dist(xx))


## get 100 bootstrap trees:
bstrees <- boot.phylo(my_nj, y,fun , B = 999,  mc.cores = 3,trees=TRUE)$trees
## get proportions of each clade:
clad <- prop.clades(my_nj, bstrees, rooted = TRUE)
## get proportions of each bipartition:
boot <- prop.clades(my_nj, bstrees)

# SVG graphics device
svg("bootstrap_plot.svg")



layout(1)
par(mar = rep(2, 4))
plot(my_nj, main = "Bipartition vs. Clade Support Values")
drawSupportOnEdges(boot)
nodelabels(clad)
legend("bottomleft", legend = c("Bipartitions", "Clades"), pch = 22,
       pt.bg = c("green", "lightblue"), pt.cex = 2.5)

# Close the graphics device
dev.off() 
```
```{r}
plot(my_nj, "phylogram")
```


```{r}
write.tree(my_nj, file = "ntw.nhx", append = FALSE,digits = 10, tree.names = FALSE)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Cmd+Option+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Cmd+Shift+K* to preview the HTML file). 

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.
