==> S22_sum_molformula_areas.ipynb <==
import numpy as np
ntw_ids= []
List=[]
with open("molformulas.txt", 'r') as f:     #Reads the Samples into a List
 
 for line in f:
     line=line.strip('\n')
     List.append(line)
     #print (line)
List.pop(0)     

List

data=[]

with open("areas_molformulas.txt", 'r') as f:     #Reads the Samples into a List
 
 for line in f:
     line=line.strip('\n')
     line=line.split('\t')
     data.append(line)
        #ListID=line.split('=')
         #ScanList.append(ListID)
     #print (line)
header=data[0]    
#data.pop(0)     

data[0]

#areas=data[0][4:116] #peaks areas
redtab=[]
for i in data:
    areas=i[4:116]
    index=i[135]
    redtab.append([areas,index])

headersarea=header[4:116]
headersarea.insert(0, "cluster")
#headersarea

merged=[]
cache=[]
for x in List: #x networkclusterid
    #cache.append(x)
    for i in data:
        if x == i[0]:
            cache.append(i[1:])
        else:
            pass
    merged.append([x,cache])
    cache=[]
            
        
#merged[0]   

merged[7]

sumlist=[]
for i in merged:
    mat = np.array(i[1]) #areas
    matflt=mat.astype(float)
    netwsum=np.sum(matflt, axis=0)
    #print(i[0])#clusterid
    sumlist.append([i[0],list(netwsum)])
    #print(netwsum)
#test=merged[0][1]#areas

#test


import numpy as np
test=merged[0][1]
#netwsum=np.empty(112)
for i in test:
    print("###")
    print(len(i))
    print((i))
    #netwsum=np.append(netwsum,i,axis=0)
mat = np.array(test)
matflt=mat.astype(float)

mat
np.shape(mat)
netwsum=np.sum(matflt, axis=0)
np.shape(netwsum)


netwsum
x=list(netwsum)

#merged[240]
x

sumlist

np.savetxt("matflt.csv", matflt, delimiter=",")


sumlist[0][1]
with open("molform-sums.csv", 'w') as f:
    for i in header:
        f.write(i)
        f.write(',')
    f.write('\n')
    for i in sumlist:
        f.write(i[0])
        f.write(',')
        for x in i[1]:
            f.write(str(x))
            f.write(',')
        f.write('\n')
