==> S28_group_for_treemap.ipynb <==
Load  summed peak areas and propagated canopus classes

import pandas as pd
df_gr=pd.read_csv("/Users/delser/Git/phd/book1.txt",sep="\t") 

df_gr

df_nset=df_gr.groupby(['canopus_enhancer_superclass','canopus_enhancer_cano_most_specific_class'])['Nset'].sum()

df_nset

headers=list(df_gr)
headers

headers.pop(0)

headers=headers[:-2]

df_sums=df_gr.groupby(['canopus_enhancer_superclass','canopus_enhancer_cano_most_specific_class'])[headers].sum()
df_sums

df_sums.to_csv("sums_groups_cano.tsv",sep="\t")
