==> S29_alpha_diversity.ipynb <==
import pandas as pd
import numpy as np


df_quant = pd.read_csv("/Users/delser/Desktop/PhD/Phytochemistry/FBMN/alltissues/FEATURE-BASED-MOLECULAR-NETWORKING-cf822b6c-15072021/alltissues15072021-quant-py-dil.csv",index_col='row ID', sep=",") # mzmine quant table


df_quant

empty=[]

from skbio.diversity import alpha_diversity
df_nrep=df_quant[["07 Nrep","Nrep03-Root_","-Nrep03-MeJ_","-Nrep03-Ctrl","-Nrep03-Caly"]].astype('int')
df_nrep_t=df_nrep.T
counts=df_nrep_t.to_numpy()
ids=["07 Nrep","Nrep03-Root_","-Nrep03-MeJ_","-Nrep03-Ctrl","-Nrep03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nnes=df_quant[["07 Nnes","Nnes03-Root_","-Nnes03-MeJ_","-Nnes03-Ctrl","-Nnes03-Caly"]].astype('int')
df_Nnes_t=df_Nnes.T
counts=df_Nnes_t.to_numpy()
ids=["07 Nnes","Nnes03-Root_","-Nnes03-MeJ_","-Nnes03-Ctrl","-Nnes03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nsto=df_quant[["07 Nsto","Nsto03-Root_","-Nsto03-MeJ_","-Nsto03-Ctrl","-Nsto03-Caly"]].astype('int')
df_Nsto_t=df_Nsto.T
counts=df_Nsto_t.to_numpy()
ids=["07 Nsto","Nsto03-Root_","-Nsto03-MeJ_","-Nsto03-Ctrl","-Nsto03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nrxs=df_quant[["07 Nrxs","Nrxs03-Root_","-Nrxs03-MeJ_","-Nrxs03-Ctrl","-Nrxs03-Caly"]].astype('int')
df_Nrxs_t=df_Nrxs.T
counts=df_Nrxs_t.to_numpy()
ids=["07 Nrxs","Nrxs03-Root_","-Nrxs03-MeJ_","-Nrxs03-Ctrl","-Nrxs03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Ntab=df_quant[["07 Ntab","-Ntab03-Root","-Ntab03-MeJ_","-Ntab03-Ctrl","-Ntab03-Caly"]].astype('int')
df_Ntab_t=df_Ntab.T
counts=df_Ntab_t.to_numpy()
ids=["07 Ntab","-Ntab03-Root","-Ntab03-MeJ_","-Ntab03-Ctrl","-Ntab03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nglu=df_quant[["08 Nglu","Nglu03-Root_","-Nglu03-MeJ_","-Nglu03-Ctrl","-Nglu03-Caly"]].astype('int')
df_Nglu_t=df_Nglu.T
counts=df_Nglu_t.to_numpy()
ids=["08 Nglu","Nglu03-Root_","-Nglu03-MeJ_","-Nglu03-Ctrl","-Nglu03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nben=df_quant[["10 Nben","-Nben03-Root","-Nben03-MeJ_","-Nben03-Ctrl","-Nben03-Caly"]].astype('int')
df_Nben_t=df_Nben.T
counts=df_Nben_t.to_numpy()
ids=["10 Nben","Nben03-Root_","-Nben03-MeJ_","-Nben03-Ctrl","-Nben03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Ncle=df_quant[["10 Ncle","-Ncle03-Root","-Ncle03-MeJ_","-Ncle03-Ctrl","-Ncle03-Caly"]].astype('int')
df_Ncle_t=df_Ncle.T
counts=df_Ncle_t.to_numpy()
ids=["10 Ncle","-Ncle03-Root","-Ncle03-MeJ_","-Ncle03-Ctrl","-Ncle03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Ngla=df_quant[["10 Ngla","-Ngla03-Root","-Ngla03-MeJ_","-Ngla03-Ctrl","-Ngla03-Caly"]].astype('int')
df_Ngla_t=df_Ngla.T
counts=df_Ngla_t.to_numpy()
ids=["10 Ngla","-Ngla03-Root","-Ngla03-MeJ_","-Ngla03-Ctrl","-Ngla03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Ngoo=df_quant[["10 Ngoo","-Ngoo03-Root","-Ngoo03-MeJ_","-Ngoo03-Ctrl","-Ngoo03-Caly"]].astype('int')
df_Ngoo_t=df_Ngoo.T
counts=df_Ngoo_t.to_numpy()
ids=["10 Ngoo","-Ngoo03-Root","-Ngoo03-MeJ_","-Ngoo03-Ctrl","-Ngoo03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Natt=df_quant[["10 Nnat","-Natt03-Root","-Natt03-MeJ_","-Natt03-Ctrl","-Natt03-Caly"]].astype('int')
df_Natt_t=df_Natt.T
counts=df_Natt_t.to_numpy()
ids=["10 Nnat","-Natt03-Root","-Natt03-MeJ_","-Natt03-Ctrl","-Natt03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nnud=df_quant[["10 Nnud","Nnud03-Root_","-Nnud03-MeJ_","-Nnud03-Ctrl","-Nnud03-Caly"]].astype('int')
df_Nnud_t=df_Nnud.T
counts=df_Nnud_t.to_numpy()
ids=["10 Nnud","Nnud03-Root_","-Nnud03-MeJ_","-Nnud03-Ctrl","-Nnud03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)


from skbio.diversity import alpha_diversity
df_Nobt=df_quant[["10 Nobt","Nobt03-Root_","-Nobt03-MeJ_","-Nobt03-Ctrl","-Nobt03-Caly"]].astype('int')
df_Nobt_t=df_Nobt.T
counts=df_Nobt_t.to_numpy()
ids=["10 Nobt","Nobt03-Root_","-Nobt03-MeJ_","-Nobt03-Ctrl","-Nobt03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Npal=df_quant[["10 Npal","-Npal03-Root","-Npal03-MeJ_","-Npal03-Ctrl","-Npal03-Caly"]].astype('int')
df_Npal_t=df_Npal.T
counts=df_Npal_t.to_numpy()
ids=["10 Npal","Npal03-Root_","-Npal03-MeJ_","-Npal03-Ctrl","-Npal03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Npan=df_quant[["10 Npan","-Npan03-Root","-Npan03-MeJ_","-Npan03-Ctrl","-Npan03-Caly"]].astype('int')
df_Npan_t=df_Npan.T
counts=df_Npan_t.to_numpy()
ids=["10 Npan","-Npan03-Root","-Npan03-MeJ_","-Npan03-Ctrl","-Npan03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Nplu=df_quant[["10 Nplu","-Nplu03-Root","-Nplu03-MeJ_","-Nplu03-Ctrl","-Nplu03-Caly"]].astype('int')
df_Nplu_t=df_Nplu.T
counts=df_Nplu_t.to_numpy()
ids=["10 Nplu","-Nplu03-Root","-Nplu03-MeJ_","-Nplu03-Ctrl","-Nplu03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Nrus=df_quant[["10 Nrus","-Nrus03-Root","-Nrus03-MeJ_","-Nrus03-Ctrl","-Nrus03-Caly","10 Nrus02mf"]].astype('int')
df_Nrus_t=df_Nrus.T
counts=df_Nrus_t.to_numpy()
ids=["10 Nrus","-Nrus03-Root","-Nrus03-MeJ_","-Nrus03-Ctrl","-Nrus03-Caly","10 Nrus02mf"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Nset=df_quant[["10 Nset","-Nset03-Root","-Nset03-MeJ_","-Nset03-Ctrl"]].astype('int')
df_Nset_t=df_Nset.T
counts=df_Nset_t.to_numpy()
ids=["10 Nset","-Nset03-Root","-Nset03-MeJ_","-Nset03-Ctrl"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Nwut=df_quant[["10 Nwut","Nwut03-Root_","-Nwut03-MeJ_","-Nwut03-Ctrl","-Nwut03-Caly"]].astype('int')
df_Nwut_t=df_Nwut.T
counts=df_Nwut_t.to_numpy()
ids=["10 Nwut","Nwut03-Root_","-Nwut03-MeJ_","-Nwut03-Ctrl","-Nwut03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

from skbio.diversity import alpha_diversity
df_Nsyl=df_quant[["10 Nsyl","Nsyl03-Root_","-Nsyl03-MeJ_","-Nsyl03-Ctrl","-Nsyl03-Caly"]].astype('int')
df_Nsyl_t=df_Nsyl.T
counts=df_Nsyl_t.to_numpy()
ids=["10 Nsyl","Nsyl03-Root_","-Nsyl03-MeJ_","-Nsyl03-Ctrl","-Nsyl03-Caly"]
adiv_obs_otus = alpha_diversity('shannon', counts, ids)
empty.append(adiv_obs_otus)
adiv_obs_otus
adiv_obs_otus = alpha_diversity('observed_otus', counts, ids)
empty.append(adiv_obs_otus)

x=pd.concat(empty, axis=0, ignore_index=False)
x

x.to_csv("alpha_diversity_nicotiana.tsv",sep="\t")
