==> S31_cosine_distance_sp_canopus.ipynb <==
import pandas as pd
df_nm=pd.read_csv("/Users/delser/ownCloud/Metabolomics-Paper-shared/NMDS.tsv",sep="\t") 
df_nm

df_gr=pd.read_csv("/Users/delser/ownCloud/Metabolomics-Paper-shared/groups.tsv",sep="\t") 
df_gr

n1=df_gr["NMDS1"].to_list()
n2=df_gr["NMDS2"].to_list()
n3=df_gr["Row.names"].to_list()
n4=df_gr["."].to_list()
sp1=df_nm["NMDS1"].to_list()
sp2=df_nm["NMDS2"].to_list()
sp3=df_nm["site"].to_list()

v_groups=[]
for v1,v2,v3,v4 in zip(n1,n2,n3,n4):
    v_groups.append([[v1,v2],v3,v4])

v_sp=[]
for v1,v2,v3 in zip(sp1,sp2,sp3):
    v_sp.append([[v1,v2],v3])

from scipy.spatial import distance
distance.cosine(v_groups[0][0], v_sp[0][0])

v_groups[0][0]

v_groups[0][2]

v_sp[0][0]

v_sp[0][1]

cosine_scores=[]
species=[]
groups_sp=[]
pval=[]
for sp in v_sp:
    print(sp[1])
    for gr in v_groups:
        species.append(sp[1])
        cosine_scores.append(distance.cosine(gr[0],sp[0]))
        groups_sp.append(gr[1])
        pval.append(gr[2])
        
        
        

len(species)



len(groups_sp)

len(cosine_scores)

df = pd.DataFrame([species,groups_sp,cosine_scores,pval])
df_transpose=df.T

df_transpose.to_csv("distance_to_canopus.tsv",sep="\t")
